requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
opencv-python==4.8.1
ultralytics==8.1.0
numpy==1.24.3
pillow==10.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
eventlet==0.33.3
